Move accessibility data fetching and parsing to unit view
